Finish I/O optimization for all scripts (finally!)

+++ catscripts/Entertainment/Video_games/Video_game_consoles/Sony/PlayStation.sh
@@ -1,17 +1,33 @@
 #!/bin/bash
 
 KEYWORDS_PS1="\bPS(| )(1\b|one)|Play(| )Station"
-KEYWORDS_PS1_EXCLUDE="Play(| )Station(| )(2|3|4|5|6|7|8|9|10|Network)|\bPSN\b|Play(| )Station(| )Portable|Play(| )Station(| )Vita|Play(| )Station(| )Store|Play(| )Station Move|Play(| )Station(| )Eye" # If this script is still being used when the PlayStation 11 comes out, I will kill myself
+SONYDIR="./catscripts/Entertainment/Video_games/Video_game_consoles/Sony"
 
-egrep -i "$KEYWORDS_PS1" newpages.txt | egrep -iv "$KEYWORDS_PS1_EXCLUDE" > PlayStation.txt
-
-PS1=$(stat --print=%s PlayStation.txt)
-
-if [ $PS1 -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="PlayStation.txt"
-  export CATNAME="PlayStation"
-  $CATEGORIZE
-fi
+  
+debug_start "PlayStation"
 
-rm PlayStation.txt
+  . $SONYDIR/PlayStation2.sh norun #KEYWORDS_PS2
+  . $SONYDIR/PlayStation3.sh norun #KEYWORDS_PLAYSTATION3
+  . $SONYDIR/PlayStation4.sh norun #KEYWORDS_PS4
+  . $SONYDIR/PlayStationNetwork.sh norun #KEYWORDS_PLAYSTATION_NETWORK
+  . $SONYDIR/PlayStationPortable.sh norun #KEYWORDS_PSP
+
+  KEYWORDS_PS1_EXCLUDE="$KEYWORDS_PS2|$KEYWORDS_PLAYSTATION3|$KEYWORDS_PS4|$KEYWORDS_PLAYSTATION_NETWORK|$KEYWORDS_PSP"
+
+  PS1=$(egrep -i "$KEYWORDS_PS1" newpages.txt | egrep -iv "$KEYWORDS_PS1_EXCLUDE")
+
+  if [ "$PS1" != "" ];
+  then
+    printf "%s" "$PS1" > PlayStation.txt
+    export CATFILE="PlayStation.txt"
+    export CATNAME="PlayStation"
+    $CATEGORIZE
+    rm PlayStation.txt
+    unset PS1
+  fi
+
+  debug_end "PlayStation"
+
+fi
