+++ catscripts/Entertainment/Video_games/Video_games_by_series/Battlefield_series/Battlefield.sh
@@ -1,24 +1,24 @@
 #!/bin/bash
 
-egrep -i 'Battlefield 1942|The Road to Rome|Secret Weapons of World War II|Battlefield series' newpages.txt >> Battlefieldseries.txt
-egrep -i 'Battlefield Vietnam' newpages.txt >> Battlefieldseries.txt
-egrep -i 'Battlefield 2 |Battlefield 2$' newpages.txt >> Battlefieldseries.txt
-egrep -i 'Battlefield 2142' newpages.txt >> Battlefieldseries.txt
-egrep -i 'Battlefield(:| ) Bad Company' newpages.txt >> Battlefieldseries.txt
-egrep -i 'Battlefield Heroes' newpages.txt >> Battlefieldseries.txt
-egrep -i 'Battlefield 1943' newpages.txt >> Battlefieldseries.txt
-egrep -i 'Battlefield Online' newpages.txt >> Battlefieldseries.txt
-egrep -i 'Battlefield Play4Free' newpages.txt >> Battlefieldseries.txt
-egrep -i 'Battlefield 3' newpages.txt >> Battlefieldseries.txt
-egrep -i 'Battlefield 4' newpages.txt >> Battlefieldseries.txt
+KEYWORDS_BATTLEFIELDSERIES="Battlefield(| )(series|1942|The(| )Road(| )to(| )Rome|Secret(| )Weapons(| )of(| )World(| )War(| )II|Vietnam|2\b|2142|(:| )Bad(| )Company|Heroes|1943|Online|Play4Free|3\b|4\b)"
 
-BATTLEFIELD=$(stat --print=%s Battlefieldseries.txt)
-
-if [ $BATTLEFIELD -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Battlefieldseries.txt"
-  export CATNAME="Battlefield series"
-  $CATEGORIZE
-fi
+  
+  debug_start "Battlefield series"
 
-rm Battlefieldseries.txt
+  BATTLEFIELD=$(egrep -i "$KEYWORDS_BATTLEFIELDSERIES" newpages.txt)
+
+  if [ "$BATTLEFIELD" != "" ];
+  then
+    printf "%s" "$BATTLEFIELD" > Battlefieldseries.txt
+    export CATFILE="Battlefieldseries.txt"
+    export CATNAME="Battlefield series"
+    $CATEGORIZE
+    rm Battlefieldseries.txt
+    unset BATTLEFIELD
+  fi
+
+  debug_end "Battlefield series"
+
+fi
