+++ catscripts/Math/Measurements/Measurements.sh
@@ -1,15 +1,27 @@
 #!/bin/bash
 
-egrep -i '(yard(|s)|feet|foot|cm|centimeters|centimetres|meters|kilometer(|s)|mile(|s)|gram(|s)|cup(|s)|ml(|\.)|mililiter(|s)|inches) is how many (inches|feet|yards|kilometer(|s)|mile(|s)|gram(|s)|meter(|s)|pound(|s)|lbs(|\.)|cup(|s)|mililiter(|s)|ml(|\.))$' newpages.txt >> MeasurementConversions.txt
-egrep -i 'equals how many meters$' newpages.txt >> MeasurementConversions.txt
+KEYWORDS_CUSTOMARY="inch|foot|feet|yard|mile|acre|fathom|ounce|quart|gallon|pound|ton|\bft\b|\byds\b|\bin\b"
+KEYWORDS_METRIC="gram|lit(re|er)|met(er|re)|\bml\b|nl\b"
+KEYWORDS_SIPREFIX="deci|centi|milli|micro|nano|femto|atto|zepto|yocto|deca|hecto|kilo|mega|giga|tera|peta|exa|zetta|yotta"
+KEYWORDS_MEASUREMENTCONVERSION="($KEYWORDS_SIPREFIX|cubic|)(| )($KEYWORDS_METRIC|$KEYWORDS_CUSTOMARY)(es|s|)(| )is(| )how(| )(many|much)(| )($KEYWORDS_SIPREFIX|cubic|)(| )($KEYWORDS_METRIC|$KEYWORDS_CUSTOMARY)|equals(| )how(| )many(| )$KEYWORDS_SIPREFIX|cubic|)(| )($KEYWORDS_METRIC|$KEYWORDS_CUSTOMARY)"
 
-CONVERSIONS=$(stat --print=%s MeasurementConversions.txt)
-
-if [ $CONVERSIONS -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="MeasurementConversions.txt"
-  export CATNAME="Measurement conversions"
-  $CATEGORIZE
-fi
 
-rm MeasurementConversions.txt
+  debug_start "Measurements"
+
+  CONVERSIONS=$(egrep -i "$KEYWORDS_MEASUREMENTCONVERSION" newpages.txt)
+
+  if [ "$CONVERSIONS" != "" ];
+  then
+    printf "%s" "$CONVERSIONS" > MeasurementConversions.txt
+    export CATFILE="MeasurementConversions.txt"
+    export CATNAME="Measurement conversions"
+    $CATEGORIZE
+    rm MeasurementConversions.txt
+    unset CONVERSIONS
+  fi
+
+  debug_end "Measurements"
+
+fi
