+++ catscripts/Science/Chemistry/Compounds/Compounds.sh
@@ -2,25 +2,37 @@
 
 ##Carbon compounds
 
-egrep -i 'Carbon(| )dioxide' newpages.txt >> CarbonDioxide.txt
-egrep -i 'Carbon(| )monoxide' newpages.txt >> CarbonMonoxide.txt
+KEYWORDS_CARBONDIOXIDE="Carbon(| )dioxide"
+KEYWORDS_CARBONMONOXIDE="Carbon(| )monoxide"
 
-CO=$(stat --print=%s CarbonMonoxide.txt)
-COTWO=$(stat --print=%s CarbonDioxide.txt)
-
-if [ $CO -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="CarbonMonoxide.txt"
-  export CATNAME="Carbon monoxide"
-  $CATEGORIZE
-fi
 
-if [ $COTWO -ne 0 ];
-then
-  export CATFILE="CarbonDioxide.txt"
-  export CATNAME="Carbon dioxide"
-  $CATEGORIZE
-fi
+  debug_start "Compounds"
 
-rm CarbonMonoxide.txt
-rm CarbonDioxide.txt
+  CO=$(egrep -i "$KEYWORDS_CARBONDIOXIDE" newpages.txt)
+  COTWO=$(egrep -i "$KEYWORDS_CARBONMONOXIDE" newpages.txt)
+
+  if [ "$CO" != "" ];
+  then
+    printf "%s" "$CO" > CarbonMonoxide.txt
+    export CATFILE="CarbonMonoxide.txt"
+    export CATNAME="Carbon monoxide"
+    $CATEGORIZE
+    rm CarbonMonoxide.txt
+    unset CO
+  fi
+
+  if [ "$COTWO" != "" ];
+  then
+    printf "%s" "$COTWO" > CarbonDioxide.txt
+    export CATFILE="CarbonDioxide.txt"
+    export CATNAME="Carbon dioxide"
+    $CATEGORIZE
+    rm CarbonDioxide.txt
+    unset COTWO
+  fi
+
+  debug_end "Compounds"
+
+fi
