+++ catscripts/Technology/Computers/Computer_programming/Computerprogramming.sh
@@ -1,23 +1,30 @@
 #!/bin/bash
 
 PROGRAMMINGDIR="./catscripts/Technology/Computers/Computer_programming"
-
-$PROGRAMMINGDIR/C++_programming/C++.sh
-$PROGRAMMINGDIR/SQL/SQL.sh
-
 KEYWORDS_COMPUTERPROGRAMMING="Computer(| )programming|Cygwin|\bMSYS\b|MATLAB"
-KEYWORDS_COMPUTERPROGRAMMING_EXCLUDE="C\+\+|SQL"
 
-egrep -i "$KEYWORDS_COMPUTERPROGRAMMING" newpages.txt | egrep -iv "$KEYWORDS_COMPUTERPROGRAMMING_EXCLUDE" > Computerprogramming.txt
-
-PROGRAMMING=$(stat --print=%s Computerprogramming.txt)
-
-if [ $PROGRAMMING -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Computerprogramming.txt"
-  export CATNAME="Computer programming"
-  $CATEGORIZE
+
+  debug_start "Computer programming"
+
+  . $PROGRAMMINGDIR/C++_programming/C++.sh #KEYWORDS_CPLUSPLUS_ALL
+  . $PROGRAMMINGDIR/SQL/SQL.sh #KEYWORDS_SQL_ALL
+
+  KEYWORDS_COMPUTERPROGRAMMING_EXCLUDE="$KEYWORDS_CPLUSPLUS_ALL|$KEYWORDS_SQL_ALL"
+
+  PROGRAMMING=$(egrep -i "$KEYWORDS_COMPUTERPROGRAMMING" newpages.txt | egrep -iv "$KEYWORDS_COMPUTERPROGRAMMING_EXCLUDE")
+
+  if [ "$PROGRAMMING" != "" ];
+  then
+    printf "%s" "$PROGRAMMING" > Computerprogramming.txt
+    export CATFILE="Computerprogramming.txt"
+    export CATNAME="Computer programming"
+    $CATEGORIZE
+    rm Computerprogramming.txt
+    unset PROGRAMMING
+  fi
+
+  debug_end "Computer programming"
+
 fi
-
-
-rm Computerprogramming.txt
